refactor: extract behavior mixins to gui/mixins/ and assemble ImmichGoGUI in gui/main_window.py
