+++ gui/__init__.py
@@ -1,0 +1,3 @@
+from gui.main_window import ImmichGoGUI
+
+__all__ = ["ImmichGoGUI"]
