+++ gui/mixins/__init__.py
@@ -1,0 +1,25 @@
+from gui.browse_dialogs import BrowseDialogsMixin
+from gui.mixins.binary_ui import BinaryUIMixin
+from gui.mixins.connection import ConnectionMixin
+from gui.mixins.diagnostics import DiagnosticsMixin
+from gui.mixins.execution import ExecutionMixin
+from gui.mixins.layout import LayoutMixin
+from gui.mixins.menu import MenuMixin
+from gui.mixins.persistence import PersistenceMixin
+from gui.mixins.profiles_ui import ProfilesUIMixin
+from gui.mixins.status import StatusMixin
+from gui.mixins.theme_mixin import ThemeMixin
+
+__all__ = [
+    "BrowseDialogsMixin",
+    "BinaryUIMixin",
+    "ConnectionMixin",
+    "DiagnosticsMixin",
+    "ExecutionMixin",
+    "LayoutMixin",
+    "MenuMixin",
+    "PersistenceMixin",
+    "ProfilesUIMixin",
+    "StatusMixin",
+    "ThemeMixin",
+]
