+++ gui/mixins/theme_mixin.py
@@ -1,0 +1,66 @@
+from PySide6.QtCore import QEvent, QSize, QTimer
+from PySide6.QtGui import QAction
+from PySide6.QtWidgets import QWidget
+
+from theme import (
+    THEME_SYSTEM,
+    apply_application_theme,
+    clear_icon_cache,
+    load_themed_icon,
+    normalize_theme_mode,
+)
+
+
+class ThemeMixin:
+    def apply_theme(self, mode=None):
+        if mode is None:
+            mode = getattr(self, "theme_mode", THEME_SYSTEM)
+        mode = normalize_theme_mode(mode)
+        self.theme_mode = mode
+        if hasattr(self, "settings"):
+            self.settings.setValue("theme_mode", mode)
+        if hasattr(self, "theme_mode_combo"):
+            self.theme_mode_combo.blockSignals(True)
+            self.theme_mode_combo.setCurrentText(mode)
+            self.theme_mode_combo.blockSignals(False)
+        resolved = apply_application_theme(mode)
+        clear_icon_cache()
+        for widget in self.findChildren(QWidget):
+            try:
+                widget.update()
+            except TypeError:
+                pass
+        self.refresh_sidebar_icons(resolved)
+        self.update()
+
+    def refresh_sidebar_icons(self, theme: str):
+        if not hasattr(self, "btn_config"):
+            return
+        nav_buttons = [
+            self.btn_config,
+            self.btn_upload,
+            self.btn_archive,
+            self.btn_stack,
+        ]
+        for btn in nav_buttons:
+            if hasattr(btn, "icon_name") and btn.icon_name:
+                btn.setIcon(load_themed_icon(btn.icon_name, theme))
+                btn.setIconSize(QSize(18, 18))
+        for action in self.findChildren(QAction):
+            if hasattr(action, "icon_name") and action.icon_name:
+                action.setIcon(load_themed_icon(action.icon_name, theme))
+
+    def on_system_theme_changed(self):
+        if getattr(self, "theme_mode", THEME_SYSTEM) == THEME_SYSTEM:
+            QTimer.singleShot(0, lambda: self.apply_theme(THEME_SYSTEM))
+
+    def _on_screen_changed(self):
+        clear_icon_cache()
+        resolved = apply_application_theme(self.theme_mode)
+        self.refresh_sidebar_icons(resolved)
+
+    def event(self, e):
+        if e.type() == QEvent.Type.ThemeChange:
+            if getattr(self, "theme_mode", THEME_SYSTEM) == THEME_SYSTEM:
+                QTimer.singleShot(0, lambda: self.apply_theme(THEME_SYSTEM))
+        return super().event(e)
